backend/requirements.txt
Flask
flask-cors
google-generativeai
Werkzeug
twilio
python-dotenv